Add audio snailCramp.mp3 && draft narrative changes for casino and each dice roll
- Recorded and then added snailcramp.mp3 to audio folder for game sound use
- Mostly drafted casino scene and what happens if you get certain dice rolls
(for example rolling a 20 is very good and you get 50 gold and can finish the game,
rolling a 1 costs you an extra 20 gold for doing so badly, etc)
- There's drafted endings you can get, running out of gold, maxing out on gold,
if you go home with enough money and if not enough money.
- I think I need to add if we run out of time we also lose. I'll do so tomorrow
as I am a rather tired now. Tomorrow will be art, sound, finishing narrative,
clean up and etc.